Consolidate UUID functionality into the UUIDFactory and make UUID1Utils test utils
This will make the API a bit more concise and backwards compatible.
The move of UUID1Utils is motivated by the fact that it's currently only used by tests,
