+++ izettle-messaging/src/main/java/com/izettle/messaging/PolledMessage.java
@@ -1,6 +1,6 @@
 package com.izettle.messaging;
 
-import com.izettle.java.uuid.UUIDFactory;
+import com.izettle.java.UUIDFactory;
 
 /**
  * Wrapper for message received from message queue. Holds both the message itself, and a handle to the message
